Avoid inline stuff in IPC
